+++ contracts/test/Mailbox.sol
@@ -42,7 +42,7 @@ contract MailBoxTest is Test {
         bytes memory data = bytes("test");
         bytes32 chainId = mailbox.chainId();
         uint256 mailboxNonce = mailbox.mailboxNonce();
-        mailbox.sendMessage(chainIdTo, to,  data);
+        mailbox.sendMessage(chainIdTo, to, data);
 
         NexusReceipt memory receipt = NexusReceipt({
             chainIdFrom: chainId,
@@ -106,7 +106,7 @@ contract MailBoxTest is Test {
         mailbox.checkVerificationOfEncoding(0, receipt, bytes32(targetChainId), value, encoding);
     }
 
-    function testSortingAlgorithm() public view{
+    function testSortingAlgorithm() public view {
         uint256 length = 5;
         bytes32[] memory chainIdTo = new bytes32[](length);
         chainIdTo[0] = bytes32(targetChainId);
@@ -115,19 +115,18 @@ contract MailBoxTest is Test {
         chainIdTo[3] = bytes32(targetChainId + 2);
         chainIdTo[4] = bytes32(targetChainId - 2);
 
-        
         address[] memory to = new address[](length);
         to[0] = address(0);
         to[1] = vm.addr(1);
         to[2] = vm.addr(2);
         to[3] = vm.addr(3);
         to[4] = vm.addr(4);
- 
-        (chainIdTo, to) = mailbox.sortWrapper(chainIdTo, to,  0, int(length-1));
+
+        (chainIdTo, to) = mailbox.sortWrapper(chainIdTo, to, 0, int256(length - 1));
 
         assertEq(chainIdTo[0], bytes32(targetChainId - 2));
-        assertEq(chainIdTo[1],bytes32(targetChainId - 1));
-        assertEq(chainIdTo[2], bytes32(targetChainId) );
+        assertEq(chainIdTo[1], bytes32(targetChainId - 1));
+        assertEq(chainIdTo[2], bytes32(targetChainId));
         assertEq(chainIdTo[3], bytes32(targetChainId + 1));
         assertEq(chainIdTo[4], bytes32(targetChainId + 2));
 
@@ -148,15 +147,14 @@ contract MailBoxTest is Test {
         chainIdTo[3] = bytes32(targetChainId + 2);
         chainIdTo[4] = bytes32(targetChainId - 2);
 
-        
         address[] memory to = new address[](length);
         to[0] = vm.addr(2);
         to[1] = vm.addr(1);
         to[2] = address(0);
         to[3] = vm.addr(3);
         to[4] = vm.addr(4);
- 
-        (chainIdTo, to) = mailbox.sortWrapper(chainIdTo, to,  0, int(length-1));
+
+        (chainIdTo, to) = mailbox.sortWrapper(chainIdTo, to, 0, int256(length - 1));
 
         address toAddr = mailbox.searchWrapper(chainIdTo, to);
         assertEq(toAddr, vm.addr(2));
